Rebuild enrollment-controller as an operator
While debugging the enrollment-controller in #20, I was inspired to consider what migrating to a proper operator might look like.

This re-implements all of the same logic/tooling for the consul and vault roles necessary on top of [kopf](https://kopf.readthedocs.io/en/stable/), which alleviates the need for tediously handling watches/streams/states.

This operator will be compatible for running side-by-side with the old controller since all operations are idempotent, but will need updates in cabotage-app to create the CabotageEnrollment resources for each application.